feat: enable a mobile to play game properly
